test: split PDF float reference scoping cases
